pyproject.toml
[tool.poetry]
name = "git-flush"
version = "0.1.4"
description = ""
authors = ["Nathan Patton <npatton@gmail.com>"]
readme = "README.md"

[tool.poetry.scripts]
git-flush = "git_flush.main:app"

[tool.poetry.dependencies]
python = "^3.9"
typer = {extras = ["all"], version = "^0.9.0"}
gitpython = "^3.1.41"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
